feat(tasks-policy): aplicando policy a las tareas del usuario.

+++ app/Http/Controllers/Api/V1/TaskController.php
@@ -2,16 +2,14 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Http\Resources\V1\TaskResource;
 use App\Models\Task;
 use App\Traits\ApiResponses;
 use Illuminate\Support\Facades\Auth;
-use PHPUnit\Framework\Attributes\Ticket;
 
-class TaskController extends Controller
+class TaskController extends ApiController
 {
     use ApiResponses;
 
@@ -45,7 +43,10 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
-        return new TaskResource($task);
+        if ($this->isAble('view', $task)){
+            return new TaskResource($task);
+        }
+        return $this->notAuthorized('You are not authorized to view this task');
     }
 
     /**
@@ -54,15 +55,18 @@ class TaskController extends Controller
     public function update(UpdateTaskRequest $request, Task $task)
     {
         // PATCH Request
-        $model = [
-            'user_id' => Auth::user()->id,
-            'category_id' => $request->input('data.relationships.category.data.id'),
-            'title' => $request->input('data.attributes.title'),
-            'description' => $request->input('data.attributes.description'),
-            'status' => $request->input('data.attributes.status'),
-        ];
-        $task->update($model);
-        return new TaskResource($task);
+        if ($this->isAble('update', $task)){
+            $model = [
+                'user_id' => Auth::user()->id,
+                'category_id' => $request->input('data.relationships.category.data.id'),
+                'title' => $request->input('data.attributes.title'),
+                'description' => $request->input('data.attributes.description'),
+                'status' => $request->input('data.attributes.status'),
+            ];
+            $task->update($model);
+            return new TaskResource($task);
+        }
+        return $this->notAuthorized('You are not authorized to update this task');
     }
 
     /**
@@ -70,7 +74,10 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
-        $task->delete();
-        return $this->ok('Task successfully deleted');
+        if ($this->isAble('delete', $task)){
+            $task->delete();
+            return $this->ok('Task successfully deleted');
+        }
+        return $this->notAuthorized('You are not authorized to delete this task');
     }
 }
